requirements.txt
# requirements.txt
# Dependências para a API Flask refinada com SQLAlchemy

# Framework web
Flask==3.0.0

# Banco de dados e ORM
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Validação e serialização de dados
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0

# Testes
pytest==7.4.4
pytest-cov==4.1.0

# Desenvolvimento (opcional)
flask-cors==4.0.0

# Logging e monitoramento (opcional)
python-json-logger==2.0.7

# Formatação de código (desenvolvimento)
black==23.12.1
flake8==7.0.0
isort==5.13.2
python-slugify
